// client/src/constant/Roles.ts
/* eslint-disable lines-around-comment */
export enum Roles {
  /**
   * USER
   */
  F1 = 1,
  /**
   * EVALUATOR
   */
  F2 = 2,
  /**
   * PRO_SKILL_SETTING
   */
  F3 = 3,
  /**
   * PRO_SKILL_APPROVAL
   */
  F4 = 4,
  /**
   * MANAGEMENT_BASIC_BEHAVIOR_SETTING
   */
  F5 = 5,
  /**
   * MANAGEMENT_EVALUATION
   */
  F6 = 6,
  /**
   * MANAGEMENT_EVALUATION_SETTING
   */
  F7 = 7,
  /**
   * MANAGEMENT_USER
   */
  F8 = 8,
  /**
   * SYSTEM_ADMIN
   */
  F9 = 9,
}